Update README.md
This repository contains Power BI product sales data. The data is in CSV format and includes the following columns:

Product Name
Product ID
Sales Date
Sales Quantity
Sales Price
The data is updated monthly.

This data can be used for a variety of purposes, such as:

Analyzing sales trends
Identifying top-selling products
Calculating sales revenue
Creating data visualizations
The data is free to use for any purpose.

Here are some additional details about the data:

The data is from 2020 to 2023.
The data is updated monthly.
The data is in CSV format.
The data is free to use for any purpose.
I hope this helps!